Moved abstractions from src/opengl to src/renderer
